refactor agent setup
