+++ openhands/core/cli.py
@@ -1,15 +1,12 @@
 import asyncio
 import logging
 import sys
-from typing import Type
 from uuid import uuid4
 
 from termcolor import colored
 
 import openhands.agenthub  # noqa F401 (we import this to get the agents registered)
 from openhands import __version__
-from openhands.controller import AgentController
-from openhands.controller.agent import Agent
 from openhands.core.config import (
     AppConfig,
     get_parser,
@@ -18,7 +15,8 @@ from openhands.core.config import (
 from openhands.core.logger import openhands_logger as logger
 from openhands.core.loop import run_agent_until_done
 from openhands.core.schema import AgentState
-from openhands.events import EventSource, EventStream, EventStreamSubscriber
+from openhands.core.setup import create_agent, create_controller, create_runtime
+from openhands.events import EventSource, EventStreamSubscriber
 from openhands.events.action import (
     Action,
     ActionConfirmationStatus,
@@ -34,11 +32,6 @@ from openhands.events.observation import (
     FileEditObservation,
     NullObservation,
 )
-from openhands.llm.llm import LLM
-from openhands.runtime import get_runtime_cls
-from openhands.runtime.base import Runtime
-from openhands.security import SecurityAnalyzer, options
-from openhands.storage import get_file_store
 
 
 def display_message(message: str):
@@ -114,39 +107,12 @@ async def main(loop):
     config = load_app_config(config_file=args.config_file)
     sid = str(uuid4())
 
-    agent_cls: Type[Agent] = Agent.get_cls(config.default_agent)
-    agent_config = config.get_agent_config(config.default_agent)
-    llm_config = config.get_llm_config_from_agent(config.default_agent)
-    agent = agent_cls(
-        llm=LLM(config=llm_config),
-        config=agent_config,
-    )
+    runtime = create_runtime(config, sid=sid, headless_mode=True)
+    await runtime.connect()
+    agent = create_agent(runtime, config)
+    controller, _ = create_controller(agent, runtime, config)
 
-    file_store = get_file_store(config.file_store, config.file_store_path)
-    event_stream = EventStream(sid, file_store)
-
-    runtime_cls = get_runtime_cls(config.runtime)
-    runtime: Runtime = runtime_cls(  # noqa: F841
-        config=config,
-        event_stream=event_stream,
-        sid=sid,
-        plugins=agent_cls.sandbox_plugins,
-        headless_mode=True,
-    )
-
-    if config.security.security_analyzer:
-        options.SecurityAnalyzers.get(
-            config.security.security_analyzer, SecurityAnalyzer
-        )(event_stream)
-
-    controller = AgentController(
-        agent=agent,
-        max_iterations=config.max_iterations,
-        max_budget_per_task=config.max_budget_per_task,
-        agent_to_llm_config=config.get_agent_to_llm_config_map(),
-        event_stream=event_stream,
-        confirmation_mode=config.security.confirmation_mode,
-    )
+    event_stream = runtime.event_stream
 
     async def prompt_for_next_task():
         # Run input() in a thread pool to avoid blocking the event loop
